feat : 여러 db의 데이터를 삭제할 수 있는 구조로 변경
