+++ eeos/src/test/java/com/blackcompany/eeos/common/DataClearExtension.java
@@ -1,7 +1,10 @@
 package com.blackcompany.eeos.common;
 
+import java.util.ArrayList;
+import java.util.List;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
+import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -10,11 +13,15 @@ public class DataClearExtension implements BeforeEachCallback {
 
 	@Override
 	public void beforeEach(ExtensionContext context) {
-		DatabaseCleaner dataCleaner = getDataCleaner(context);
-		dataCleaner.clear();
+		ApplicationContext appContext = SpringExtension.getApplicationContext(context);
+		List<DatabaseCleaner> dataCleaners = getDataCleaners(appContext);
+
+		for (DatabaseCleaner cleaner : dataCleaners) {
+			cleaner.clear();
+		}
 	}
 
-	private DatabaseCleaner getDataCleaner(final ExtensionContext extensionContext) {
-		return SpringExtension.getApplicationContext(extensionContext).getBean(DatabaseCleaner.class);
+	private List<DatabaseCleaner> getDataCleaners(ApplicationContext appContext) {
+		return new ArrayList<>(appContext.getBeansOfType(DatabaseCleaner.class).values());
 	}
 }
